protocol: refactor and separate epoch/updater tests
